+++ crates/roc_host/src/lib.rs
@@ -1102,132 +1102,17 @@ pub struct CommandOutput {
 }
 
 #[no_mangle]
-pub extern "C" fn roc_fx_commandStatus(roc_cmd: &Command) -> RocResult<(), RocList<u8>> {
-    let args = roc_cmd.args.into_iter().map(|arg| arg.as_str());
-    let num_envs = roc_cmd.envs.len() / 2;
-    let flat_envs = &roc_cmd.envs;
-
-    // Environment vairables must be passed in key=value pairs
-    assert_eq!(flat_envs.len() % 2, 0);
-
-    let mut envs = Vec::with_capacity(num_envs);
-    for chunk in flat_envs.chunks(2) {
-        let key = chunk[0].as_str();
-        let value = chunk[1].as_str();
-        envs.push((key, value));
-    }
-
-    // Create command
-    let mut cmd = std::process::Command::new(roc_cmd.program.as_str());
-
-    // Set arguments
-    cmd.args(args);
-
-    // Clear environment variables if cmd.clearEnvs set
-    // otherwise inherit environment variables if cmd.clearEnvs is not set
-    if roc_cmd.clearEnvs {
-        cmd.env_clear();
-    };
-
-    // Set environment variables
-    cmd.envs(envs);
-
-    match cmd.status() {
-        Ok(status) => {
-            if status.success() {
-                RocResult::ok(())
-            } else {
-                match status.code() {
-                    Some(code) => commandStatusErrorCode(code),
-                    None => {
-                        // If no exit code is returned, the process was terminated by a signal.
-                        commandStatusKilledBySignal()
-                    }
-                }
-            }
-        }
-        Err(err) => commandStatusOtherError(err),
-    }
-}
-
-fn commandStatusKilledBySignal() -> RocResult<(), RocList<u8>> {
-    let mut error_bytes = Vec::new();
-    error_bytes.extend([b'K', b'S']);
-    let error = RocList::from(error_bytes.as_slice());
-    RocResult::err(error)
-}
-
-fn commandStatusErrorCode(code: i32) -> RocResult<(), RocList<u8>> {
-    let mut error_bytes = Vec::new();
-    error_bytes.extend([b'E', b'C']);
-    error_bytes.extend(code.to_ne_bytes()); // use NATIVE ENDIANNESS
-    let error = RocList::from(error_bytes.as_slice()); //RocList::from([b'E',b'C'].extend(code.to_le_bytes()));
-    RocResult::err(error)
-}
-
-fn commandStatusOtherError(err: std::io::Error) -> RocResult<(), RocList<u8>> {
-    let error = RocList::from(format!("{:?}", err).as_bytes());
-    RocResult::err(error)
+pub extern "C" fn roc_fx_commandStatus(
+    roc_cmd: &roc_command::Command,
+) -> RocResult<(), RocList<u8>> {
+    roc_command::command_status(roc_cmd)
 }
 
 #[no_mangle]
-pub extern "C" fn roc_fx_commandOutput(roc_cmd: &Command) -> CommandOutput {
-    let args = roc_cmd.args.into_iter().map(|arg| arg.as_str());
-    let num_envs = roc_cmd.envs.len() / 2;
-    let flat_envs = &roc_cmd.envs;
-
-    // Environment vairables must be passed in key=value pairs
-    assert_eq!(flat_envs.len() % 2, 0);
-
-    let mut envs = Vec::with_capacity(num_envs);
-    for chunk in flat_envs.chunks(2) {
-        let key = chunk[0].as_str();
-        let value = chunk[1].as_str();
-        envs.push((key, value));
-    }
-
-    // Create command
-    let mut cmd = std::process::Command::new(roc_cmd.program.as_str());
-
-    // Set arguments
-    cmd.args(args);
-
-    // Clear environment variables if cmd.clearEnvs set
-    // otherwise inherit environment variables if cmd.clearEnvs is not set
-    if roc_cmd.clearEnvs {
-        cmd.env_clear();
-    };
-
-    // Set environment variables
-    cmd.envs(envs);
-
-    match cmd.output() {
-        Ok(output) => {
-            // Status of the child process, successful/exit code/killed by signal
-            let status = if output.status.success() {
-                RocResult::ok(())
-            } else {
-                match output.status.code() {
-                    Some(code) => commandStatusErrorCode(code),
-                    None => {
-                        // If no exit code is returned, the process was terminated by a signal.
-                        commandStatusKilledBySignal()
-                    }
-                }
-            };
-
-            CommandOutput {
-                status,
-                stdout: RocList::from(&output.stdout[..]),
-                stderr: RocList::from(&output.stderr[..]),
-            }
-        }
-        Err(err) => CommandOutput {
-            status: commandStatusOtherError(err),
-            stdout: RocList::empty(),
-            stderr: RocList::empty(),
-        },
-    }
+pub extern "C" fn roc_fx_commandOutput(
+    roc_cmd: &roc_command::Command,
+) -> roc_command::CommandOutput {
+    roc_command::command_output(roc_cmd)
 }
 
 #[no_mangle]
